# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个基于 Go 语言构建的企业资源规划（ERP）后端系统，使用 JWT 进行用户认证，PostgreSQL 作为数据库。API接口按功能模块分组，提供统一的路径结构。",
        "title": "ERP 后端系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/product": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分页获取商品列表，支持多种筛选条件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品管理"
                ],
                "summary": "获取商品列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量，默认10",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "商品名称搜索（模糊匹配）",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "SKU精确搜索",
                        "name": "sku",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "商品编码搜索（模糊匹配）",
                        "name": "product_code",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "货源ID筛选",
                        "name": "source_id",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "最低价格",
                        "name": "min_price",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "最高价格",
                        "name": "max_price",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否优惠筛选",
                        "name": "is_discounted",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否启用筛选",
                        "name": "is_enabled",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "颜色筛选（可多选）",
                        "name": "colors",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "发货时间筛选（模糊匹配）",
                        "name": "shipping_time",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "id",
                            "name",
                            "sku",
                            "product_code",
                            "price",
                            "discount_price",
                            "cost_price",
                            "is_discounted",
                            "is_enabled",
                            "shipping_time",
                            "created_at",
                            "updated_at"
                        ],
                        "type": "string",
                        "description": "排序字段: id, name, sku, product_code, price, discount_price, cost_price, is_discounted, is_enabled, shipping_time, created_at, updated_at",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "排序方向: asc, desc",
                        "name": "order_dir",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/definitions/model.Product"
                                                    }
                                                },
                                                "page": {
                                                    "type": "integer"
                                                },
                                                "page_size": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "total_pages": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新的商品信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品管理"
                ],
                "summary": "创建商品",
                "parameters": [
                    {
                        "description": "商品信息",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Product"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/product/code/{code}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "通过SKU获取商品详情，并记录查询历史",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "商品管理"
                ],
                "summary": "通过SKU获取商品",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商品SKU",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Product"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "商品不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入 \"Bearer \" 加上 JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}